# Optimize game 1