--- a/README.md
+++ b/README.md
@@ -1,1 +1,17 @@
 # zombie-survival
+
+## how to run "on desktop"
+
+python main.py
+
+##  how to run on web
+
+You need use pygbag. It builds and set up a server for you.
+
+By default it starts webserver at port 8080. goto localhost:8080
+
+pygbag seem to require the entrypoint to be called main.py. Do not rename that file.
+
+
+
+