--- a/tilemap.py
+++ b/tilemap.py
@@ -1,6 +1,7 @@
 import pygame
 import json
 from spritesheet import Sprite
+from main import scene_position_to_view_port_position
 
 def get_json(filename):
     with open(filename, 'r') as f:
@@ -12,7 +13,6 @@ def get_obj(obj, name):
             return o
 
 def get_tile_imgs(game_map, size, obj, img_name):
-    IMGs = []
     allTiles = get_obj(game_map, obj)
     tileimg = Sprite(img_name)
 
@@ -24,20 +24,17 @@ def get_tile_imgs(game_map, size, obj, img_name):
         if tp[0] > max_x: max_x = tp[0]
         if tp[1] < min_y: min_y = tp[1]
         if tp[1] > max_y: max_y = tp[1]
-
+    
     layer_size = ( max_x - min_x ) + size, ( max_y - min_y ) + size
     layer = pygame.Surface( layer_size )
-    layer.fill( (144,244,200) )
+    layer.set_colorkey((0, 0, 0))
     for tile in allTiles["gridTiles"]:
 
         tile_surface = pygame.transform.flip(tileimg.get_sprite(tile["src"], size), tile['f'], False)
         tile_position = tile["px"]
-        IMGs.append([tile_surface, tile_position])
         layer.blit( tile_surface, tile_position )
     
-
-
-    return IMGs, layer
+    return layer
 
 
 def drawing_layers(win, lis):
@@ -68,7 +65,7 @@ class Tiledmap:
         #   the first 2 parameters will be the same for all
         #   the third parameter will the layer's name in LDtk
         #   the fourth parameter is the image's path
-        self.TileL1, self.layer1 = get_tile_imgs(self.game_map, self.size, "Tiles", "map/Terrain_and_Props.png")
+        self.layer1 = get_tile_imgs(self.game_map, self.size, "Tiles", "map/Terrain_and_Props.png")
 
         self.csvMap = csvMAP = get_obj(self.game_map, "Grid_set")
         x = 0
@@ -88,31 +85,14 @@ class Tiledmap:
                 cx = 0
                 cy += 1
 
-    # old: (remove later)
-    def _draw(self, win):
-        from main import scene_position_to_view_port_position
-        for tile in self.TileL2 + self.TileL1:
-            # NOTE/TODO should we check for screen_check here?
-            projected_position = scene_position_to_view_port_position( (tile[1][0],tile[1][1]) )  
-            win.blit(tile[0], projected_position )
-
-
-
     def get_cell( self, cell_position ):
         #print(sorted(self.cells.keys()))
         return self.cells.get( cell_position, 0)
     
     def draw(self, pl, win):
 
-
-        #drawing_layers(win, self.TileL1)
-
-        from main import scene_position_to_view_port_position
-
         pp = projected_position = scene_position_to_view_port_position( (0,0) )  
 
-        win.blit(self.layer1,pp)
-
-
-
         pl.draw(win)
+
+        win.blit(self.layer1,pp)