I done some optimizations.

```              
                Desktop    Web
This version:   208 FPS    54 FPS
Old version:    175 FPS    40 FPS
```

The new code make use of the idea to blit a big background instead of many tiles.

We blit all the tiles to a background surface. Even though the background can be big it should not be 
a problem since pygame doesn't draw outside the game screen. It just draw the intersection of the two surfaces.

Yes, this solution is more memory heavy, hopefully that is ok