@coppermouse 2023-02-07
Another comment: Now player is being drawn onto the background, not behind. I can fix so we have layers of backgrounds later.